Move graphics scripts to folders by methodology or variant.
