+++ relecov_dashboard/utils/variants/graphics/needle_mutation_graph_by_sample.py
@@ -11,8 +11,7 @@ from django_plotly_dash import DjangoDash
 # Local imports
 import relecov_core.utils.handling_variant
 
-# TODO: mv to graphics
-# TODO: rename to create_needle_graph()
+# FIXME: This file is not accessed.
 def create_needle_plot_graph_mutation_by_sample(sample_name, mdata):
     sample_list = [2018185, 210067]
     app = DjangoDash("needlePlotBySample")
